+++ include/executor.h
@@ -4,6 +4,7 @@
 #include "parser.h"
 
 int handle_internal_command(struct command cmd);
+
 int execute_commands(struct command* cmds, int num_cmds);
 
 #endif
